requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
openai==0.28.1
python-dotenv==1.0.1
geckodriver-autoinstaller==0.1.0 